requirements.txt
streamlit
transformers
torchaudio
openai-whisper
moviepy
pandas
